requirements.txt
django
django-rest-framework
psycopg2
gunicorn
whitenoise
django-cors-headers
django-silk
pillow
python-decouple
dj-database-url
django-allauth
django-allauth[socialaccount]
django-phonenumber-field[phonenumbers]
django-countries
# geopy
# pycountry
drf-spectacular
drf-spectacular[sidecar]
django-filter
python-dotenv
